returns pointer to newly allocated space in memory

+++ 0x0B-malloc_free/1-strdup.c
@@ -9,9 +9,9 @@
  */
 char *_strdup(char *str)
 {
-	size_t len = strlen(str) + 1;
+	int len = strlen(str) + 1;
 	char *mem_al;
-	
+
 	if (str == NULL)
 	{
 		return (NULL);
